rename some scripts and update logic of TemporaryFileManager

+++ scripts/temporary_file_manager.py
@@ -8,12 +8,16 @@ class TemporaryFileManager:
 		self.name = tmp_file_name
 
 	def delete_tmp_file(self):
-		os.remove(self.name)
+		if os.path.exists(self.name):
+			os.remove(self.name)
 
 	def register_at_exit(self):
 		atexit.register(self.delete_tmp_file)
 
 	@classmethod
 	def create(cls, suffix, tmp_dir=None):
-		tmp_file_name = tempfile.NamedTemporaryFile(dir=tmp_dir, suffix=suffix).name
-		return cls(tmp_file_name)
+		tmp_file = tempfile.NamedTemporaryFile(dir=tmp_dir, suffix=suffix, delete=False)
+		tmp_file.close()  # Close the file to ensure it's created on disk
+		tmp_file_manager = cls(tmp_file.name)
+		tmp_file_manager.register_at_exit()
+		return tmp_file_manager
